made minor comment suggestions and changed item of interface returning to double

+++ src/Command/Backward.java
@@ -21,7 +21,7 @@ public class Backward implements Command {
     }
 
     @Override
-    public int returnValue() {
-        return (int)steps;
+    public double returnValue() {
+        return (double)steps;
     }
 }
